rename package to vyosnetworks_vyos
