+++ nodes/vyosnetworks_vyos/vyos.config.boot
